Move ToolCommand into submodule
