+++ crates/volta-core/src/tool/command/binary.rs
