+++ crates/volta-core/src/tool/command/npm.rs
