+++ crates/volta-core/src/tool/command/npx.rs
